Исправления в Template и Main
