+++ src/components/NewCollection/NewCollection.module.scss
@@ -1,8 +1,3 @@
-@font-face {
-    font-family: "Raleway";
-    src: url("../../../assets/fonts/Montserrat-VariableFont_wght.ttf");
-}
-
 $primary-color: #6E9C9F;
 $font-family: "Raleway";
 $font-weight-regular: 450;
@@ -77,52 +72,53 @@ $font-weight-regular: 450;
         }
     }
 
-        .imageWrapper {
-            position: relative;
-            overflow: hidden;
-            display: inline-block;
+    .imageWrapper {
+        position: relative;
+        overflow: hidden;
+        display: inline-block;
+
+        .productPhoto,
+        .productPhoto2,
+        .productPhoto3 {
+            transition: opacity 0.3s ease;
+            width: 100%;
+            height: 100%;
+        }
+
+        .hoverButton {
+            position: absolute;
+            top: 50%;
+            left: 50%;
+            transform: translate(-50%, -50%);
+            background-color: transparent;
+            border: none;
+            cursor: pointer;
+            opacity: 0;
+        }
+
+        &:hover {
 
             .productPhoto,
             .productPhoto2,
             .productPhoto3 {
-                transition: opacity 0.3s ease;
-                width: 100%;
-                height: 100%;
+                opacity: 0.3;
             }
 
             .hoverButton {
-                position: absolute;
-                top: 50%;
-                left: 50%;
-                transform: translate(-50%, -50%);
-                background-color: transparent;
-                border: none;
-                cursor: pointer;
-                opacity: 0;
+                visibility: visible;
+                opacity: 1;
             }
-            &:hover {
 
-                .productPhoto,
-                .productPhoto2,
-                .productPhoto3 {
-                    opacity: 0.3;
-                }
-
-                .hoverButton {
-                    visibility: visible;
-                    opacity: 1;
-                }
-
-                &::after {
-                    content: "";
-                    position: absolute;
-                    top: 0;
-                    left: 0;
-                    right: 0;
-                    bottom: 0;
-                    background-color: #6E9C9F;
-                    z-index: -1;
-                }
+            &::after {
+                content: "";
+                position: absolute;
+                top: 0;
+                left: 0;
+                right: 0;
+                bottom: 0;
+                background-color: #6E9C9F;
+                z-index: -1;
             }
         }
     }
+}
