+++ src/components/Template/Template.jsx
@@ -1,4 +1,3 @@
-import { NewCollection } from "./NewCollection";
 import styles from "./Template.module.scss";
 
 export function Template() {
